+++ src/remotes/remotefactory.ts
@@ -3,6 +3,7 @@ import * as assert from 'assert';
 import { ZesaruxExtRemote } from './zesarux/zesaruxextremote';
 import { RemoteClass } from './remoteclass';
 import {ZxNextUsbSerialRemote} from './zxnext/zxnextusbserialremote';
+import {ZxSimulatorRemote} from './zxsimulator/zxsimremote';
 
 
 
@@ -12,16 +13,19 @@ import {ZxNextUsbSerialRemote} from './zxnext/zxnextusbserialremote';
 export class RemoteFactory {
 	/**
 	 * Factory method to create an emulator.
-	 * @param remoteType 'zrcp' or 'zxnext'. For 'zrcp' always the ZesaruxExtEmulator is created.
+	 * @param remoteType 'zrcp', 'serial' or 'zxsim'. For 'zrcp' always the ZesaruxExtEmulator is created.
 	 * It will fallback to Zesarux if no ZesaruxExt is connected.
 	 */
 	public static createRemote(remoteType: string) {
 		switch (remoteType) {
-			case 'zrcp':
+			case 'zrcp':	// ZEsarUX Remote Control Protocol
 				RemoteFactory.setEmulator(new ZesaruxExtRemote());
 				break;
-			case 'zxnext':
+			case 'serial':	// USB/serial connection
 				RemoteFactory.setEmulator(new ZxNextUsbSerialRemote());
+				break;
+			case 'zxsim':	// Simulator
+				RemoteFactory.setEmulator(new ZxSimulatorRemote());
 				break;
 			case 'mame':
 				assert(false);	// needs to be implemented